+++ src/coloring/set_color_matrix.c
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 19:04:24 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/20 23:21:46 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/21 17:41:56 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,11 +37,13 @@ t_colors	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 	int			y;
 
 //! ~~ Testing started ~~~~~~~~~~
-/*	void	*gnrl_ptr;
+//	/*	
+	void	*gnrl_ptr;
 	void	*win_ptr;
 
 	gnrl_ptr = mlx_init();
-	win_ptr = mlx_new_window(gnrl_ptr, HORIZ_RESOLUTION, RESOLUTION, "miniRT test");*/
+	win_ptr = mlx_new_window(gnrl_ptr, HORIZ_RESOLUTION, RESOLUTION, "miniRT test");
+//	*/
 //! ~~~ Testing ended ~~~~~~~~~~~
 
 	c_matrix = malloc(sizeof(t_coords *) * HORIZ_RESOLUTION);
@@ -61,9 +63,9 @@ t_colors	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 			{
 
 			//! ~~ Testing started ~~~~~~~~~~
-/*
+//	/*
 				mlx_pixel_put (gnrl_ptr, win_ptr, x, y, 16777215);
-*/
+//	*/
 			//! ~~~ Testing ended ~~~~~~~~~~~
 
 				//TODO	Here we'll calculate collision point & get its color.
@@ -74,9 +76,9 @@ t_colors	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 	}
 
 //! ~~ Testing started ~~~~~~~~~~
-/*
+//	/*
 	mlx_loop(gnrl_ptr);
-*/
+//	*/
 //! ~~~ Testing ended ~~~~~~~~~~~
 
 	return (c_matrix);